refactor(ocr): move the parser registry out of the package __init__
`_PYMUPDF_AVAILABLE` tracked exactly what `"pymupdf" in _PARSERS` already
answers, and its sole reader was `default_parser_name`. The membership test
replaces it, so the registry dict is the only record of what is installed.

The registry itself moves to `parsers/registry.py`, leaving `__init__.py` empty:
it held executable code, and two of its imports sat mid-file because the
registration statements around them needed ordering. Both imports are now at the
top of a plain module, the optional pymupdf one still behind its `try`, and the
seven call sites import from `parsers.registry`.
